Refactor options and files resolve functionality to allow Go invocations
The purpose of this change is to allow other Go projects to make use of the `ko resolve` feature programmatically. In order to do so, we've moved all the model options to a common module (`parameters`) and kept all Cobra parsing and CLI related code in `commands`.
Besides this, we have made public the MakeBuilder, MakePublisher and ResolveFilesToWriter functions. These will be used by other projects in order to perform the desired functionality.

This change does not introduce any improvements or regression beyond the new structuring and layout.
